v4.3.3
FIXED:
- A bug, where enchantment tick interval works incorrect.

+++ API/src/main/java/su/nightexpress/excellentenchants/api/ConfigBridge.java
@@ -10,7 +10,7 @@ public class ConfigBridge {
     private static boolean globalDistTradeEquipment;
     private static boolean globalDistRandomLoot;
 
-    public static int getEnchantsTickInterval() {
+    public static long getEnchantsTickInterval() {
         return enchantsTickInterval;
     }
